+++ src/main/java/com/project/domain/notification/infra/messaging/NotificationKafkaConsumer.java
@@ -10,7 +10,7 @@ import com.dabom.messaging.kafka.event.KafkaEventMessageSupport;
 import com.dabom.messaging.kafka.event.dto.EventEnvelope;
 import com.dabom.messaging.kafka.event.dto.notification.NotificationPayload;
 import com.fasterxml.jackson.core.type.TypeReference;
-import com.project.domain.notification.service.NotificationService;
+import com.project.domain.notification.service.NotificationEventService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -21,7 +21,7 @@ import lombok.extern.slf4j.Slf4j;
 public class NotificationKafkaConsumer {
 
     private final KafkaEventMessageSupport kafkaEventMessageSupport;
-    private final NotificationService notificationService;
+    private final NotificationEventService notificationEventService;
 
     @KafkaListener(topics = KafkaTopics.NOTIFICATION, groupId = "${spring.kafka.consumer.group-id}")
     public void consume(ConsumerRecord<String, String> consumerRecord) {
@@ -34,7 +34,7 @@ public class NotificationKafkaConsumer {
                             "Notification 수신 type={}, familyId={}",
                             envelope.payload().type(),
                             envelope.payload().familyId());
-                    notificationService.handleNotificationEvent(
+                    notificationEventService.handleNotificationEvent(
                             envelope.payload(), envelope.timestamp());
                 });
     }
