+++ src/main/java/com/project/domain/notification/service/NotificationService.java
@@ -1,15 +1,11 @@
 package com.project.domain.notification.service;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
-import com.dabom.messaging.kafka.event.dto.notification.NotificationPayload;
 import com.dabom.messaging.kafka.event.dto.notification.NotificationType;
 import com.project.domain.notification.dto.NotificationSlice;
 
 public interface NotificationService {
-
-    void handleNotificationEvent(NotificationPayload payload, LocalDateTime sentAt);
 
     NotificationSlice getNotifications(
             Long customerId, String cursor, int size, Boolean isRead, List<NotificationType> types);
